Refactored names of nodefilters
subscript.scripts.nfilters provides some of the most used code in this
repository. This refactor is intended to allow for more convenience when
importing nodefilter modules. I have found it is more convenient to
import all nodefilter at once, for example:

```from subscript.scripts.nfilters as nf```

This way nodefilters should be able to be accessed in a single import
, for example:

```nf.subhalos(None)```

shows the level of convenience this allows for.
This refactor is designed to encourage this syntax.
This commit is non breaking, old function names will continue to be
supported.
Added testing for old functions in

test/test_nfilters_legacy.py